+++ docs/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Moved CHANGELOG from `.github` to `docs`
     - `actions_bootstrap.ps1` - bumped module versions to latest
     - All Infra/Infrastructure references changed to Integration
+    - Removed all test case uses of `Assert-MockCalled`
 - Misc
     - Updated README to reference new badge urls for Github actions
     - Updated `settings.json` for tab requirements to support Readthedocs